Refactoring

+++ Project/ZL.Unity/Assets/ZL/CS/Scripts/MathFEx.cs
@@ -2,6 +2,9 @@ using System;
 
 namespace ZL.CS
 {
+    /// <summary>
+    /// Extenstions of MathF
+    /// </summary>
     public static partial class MathFEx
     {
         /// <summary>
@@ -29,16 +32,27 @@ namespace ZL.CS
         /// </summary>
         public const float OneOver255 = 0.003921569f;
 
-        public static float PercentToDecibel(int value)
+        /// <summary>
+        /// Convert percentage value (0 to 100) to decibel.
+        /// </summary>
+        public static float PercentToDecibel(float value)
         {
             return LinearToDecibel(value * 0.01f);
         }
 
-        public static int DecibelToPercent(float value)
+        /// <summary>
+        /// Convert decibel value to percentage (0 to 100).
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
+        public static float DecibelToPercent(float value)
         {
-            return (int)DecibelToLinear(value) * 100;
+            return DecibelToLinear(value) * 100f;
         }
 
+        /// <summary>
+        /// Convert linear value (0.0 to 1.0) to decibel.
+        /// </summary>
         public static float LinearToDecibel(float value)
         {
             if (value <= 0f)
@@ -49,6 +63,9 @@ namespace ZL.CS
             return MathF.Log10(value) * 20f;
         }
 
+        /// <summary>
+        /// Convert decibel value to linear (0.0 to 1.0).
+        /// </summary>
         public static float DecibelToLinear(float value)
         {
             if (value <= -80f)
